+++ R/cor_test.R
@@ -4,68 +4,68 @@
 #'
 #' @param data A data frame.
 #' @param x,y Names of two variables present in the data.
-#' @param ci Confidence/Credible Interval level. If \code{"default"}, then it is
-#'   set to 0.95 (95\% CI).
+#' @param ci Confidence/Credible Interval level. If `"default"`, then it is
+#'   set to `0.95` (`95%` CI).
 #' @param method A character string indicating which correlation coefficient is
-#'   to be used for the test. One of \code{"pearson"} (default),
-#'   \code{"kendall"}, \code{"spearman"} (but see also the \code{robust} argument), \code{"biserial"},
-#'   \code{"polychoric"}, \code{"tetrachoric"}, \code{"biweight"},
-#'   \code{"distance"}, \code{"percentage"} (for percentage bend correlation),
-#'   \code{"blomqvist"} (for Blomqvist's coefficient), \code{"hoeffding"} (for
-#'   Hoeffding's D), \code{"gamma"}, \code{"gaussian"} (for Gaussian Rank
-#'   correlation) or \code{"shepherd"} (for Shepherd's Pi correlation). Setting
-#'   \code{"auto"} will attempt at selecting the most relevant method
+#'   to be used for the test. One of `"pearson"` (default),
+#'   `"kendall"`, `"spearman"` (but see also the `robust` argument), `"biserial"`,
+#'   `"polychoric"`, `"tetrachoric"`, `"biweight"`,
+#'   `"distance"`, `"percentage"` (for percentage bend correlation),
+#'   `"blomqvist"` (for Blomqvist's coefficient), `"hoeffding"` (for
+#'   Hoeffding's D), `"gamma"`, `"gaussian"` (for Gaussian Rank
+#'   correlation) or `"shepherd"` (for Shepherd's Pi correlation). Setting
+#'   `"auto"` will attempt at selecting the most relevant method
 #'   (polychoric when ordinal factors involved, tetrachoric when dichotomous
 #'   factors involved, point-biserial if one dichotomous and one continuous and
 #'   pearson otherwise).
 #' @param bayesian,partial_bayesian If TRUE, will run the correlations under a
 #'   Bayesian framework. Note that for partial correlations, you will also need
-#'   to set \code{partial_bayesian} to \code{TRUE} to obtain "full" Bayesian
+#'   to set `partial_bayesian` to `TRUE` to obtain "full" Bayesian
 #'   partial correlations. Otherwise, you will obtain pseudo-Bayesian partial
 #'   correlations (i.e., Bayesian correlation based on frequentist
 #'   partialization).
-#' @param include_factors If \code{TRUE}, the factors are kept and eventually
+#' @param include_factors If `TRUE`, the factors are kept and eventually
 #'   converted to numeric or used as random effects (depending of
-#'   \code{multilevel}). If \code{FALSE}, factors are removed upfront.
-#' @param partial Can be \code{TRUE} or \code{"semi"} for partial and
+#'   `multilevel`). If `FALSE`, factors are removed upfront.
+#' @param partial Can be `TRUE` or `"semi"` for partial and
 #'   semi-partial correlations, respectively.
-#' @inheritParams effectsize::adjust
+#' @inheritParams datawizard::adjust
 #' @param bayesian_prior For the prior argument, several named values are
-#'   recognized: \code{"medium.narrow"}, \code{"medium"}, \code{"wide"}, and
-#'   \code{"ultrawide"}. These correspond to scale values of \code{1/sqrt(27)},
-#'   \code{1/3}, \code{1/sqrt(3)} and \code{1}, respectively. See the
-#'   \code{BayesFactor::correlationBF} function.
+#'   recognized: `"medium.narrow"`, `"medium"`, `"wide"`, and
+#'   `"ultrawide"`. These correspond to scale values of `1/sqrt(27)`,
+#'   `1/3`, `1/sqrt(3)` and `1`, respectively. See the
+#'   `BayesFactor::correlationBF` function.
 #' @param bayesian_ci_method,bayesian_test See arguments in
-#'   \code{\link[=parameters]{model_parameters}} for \code{BayesFactor} tests.
-#' @param ranktransform If \code{TRUE}, will rank-transform the variables prior to
+#'   [`model_parameters()`][parameters] for `BayesFactor` tests.
+#' @param ranktransform If `TRUE`, will rank-transform the variables prior to
 #'   estimating the correlation, which is one way of making the analysis more
 #'   resistant to extreme values (outliers). Note that, for instance, a Pearson's
 #'   correlation on rank-transformed data is equivalent to a Spearman's rank
-#'   correlation. Thus, using \code{robust=TRUE} and \code{method="spearman"} is
+#'   correlation. Thus, using `robust=TRUE` and `method="spearman"` is
 #'   redundant. Nonetheless, it is an easy option to increase the robustness of the
 #'   correlation as well as flexible way to obtain Bayesian or multilevel
 #'   Spearman-like rank correlations.
-#' @param robust Old name for \code{ranktransform}. Will be removed in subsequent
-#'   versions, so better to use \code{ranktransform} which is more explicit about
+#' @param robust Old name for `ranktransform`. Will be removed in subsequent
+#'   versions, so better to use `ranktransform` which is more explicit about
 #'   what it does.
 #' @param winsorize Another way of making the correlation more "robust" (i.e.,
-#'   limiting the impact of extreme values). Can be either \code{FALSE} or a
-#'   number between 0 and 1 (e.g., \code{0.2}) that corresponds to the desired
-#'   threshold. See the \code{\link[=winsorize]{winsorize()}} function for more details.
+#'   limiting the impact of extreme values). Can be either `FALSE` or a
+#'   number between 0 and 1 (e.g., `0.2`) that corresponds to the desired
+#'   threshold. See the [`winsorize()`][winsorize] function for more details.
 #' @param verbose Toggle warnings.
-#' @param ... Additional arguments (e.g., \code{alternative}) to be passed to
-#'   other methods. See \code{stats::cor.test} for further details.
+#' @param ... Additional arguments (e.g., `alternative`) to be passed to
+#'   other methods. See `stats::cor.test` for further details.
 #'
 #'
 #' @inherit correlation details
-#'
-#'
 #'
 #' @examples
 #' library(correlation)
 #'
 #' cor_test(iris, "Sepal.Length", "Sepal.Width")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "spearman")
+#'
+#' \dontrun{
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "kendall")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "biweight")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "distance")
@@ -82,6 +82,13 @@
 #' if (require("BayesFactor", quietly = TRUE)) {
 #'   cor_test(iris, "Sepal.Length", "Sepal.Width", bayesian = TRUE)
 #' }
+#'
+#' # Robust (these two are equivalent)
+#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "spearman")
+#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "pearson", ranktransform = TRUE)
+#'
+#' # Winsorized
+#' cor_test(iris, "Sepal.Length", "Sepal.Width", winsorize = 0.2)
 #'
 #' # Tetrachoric
 #' if (require("psych", quietly = TRUE)) {
@@ -102,21 +109,11 @@
 #'   cor_test(data, "Sepal.Width", "Sepal.Length_ordinal", method = "polychoric")
 #' }
 #'
-#' # Robust (these two are equivalent)
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "spearman")
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "pearson", ranktransform = TRUE)
-#'
-#' # Winsorized
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", winsorize = 0.2)
-#' \dontrun{
 #' # Partial
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", partial = TRUE)
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", multilevel = TRUE)
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", partial_bayesian = TRUE)
 #' }
-#' @importFrom effectsize adjust ranktransform
-#' @importFrom parameters data_to_numeric
-#' @importFrom stats complete.cases na.omit sd
 #' @export
 cor_test <- function(data,
                      x,
@@ -143,7 +140,7 @@ cor_test <- function(data,
     ranktransform <- robust
   }
 
-  # Sanity checks
+  # valid matrix checks
   if (!x %in% names(data) | !y %in% names(data)) {
     stop("The names you entered for x and y are not available in the dataset. Make sure there are no typos!")
   }
@@ -153,13 +150,21 @@ cor_test <- function(data,
 
   # Make sure factor is no factor
   if (!method %in% c("tetra", "tetrachoric", "poly", "polychoric")) {
-    data[c(x, y)] <- parameters::data_to_numeric(data[c(x, y)], dummy_factors = FALSE)
+    data[c(x, y)] <- datawizard::data_to_numeric(data[c(x, y)], dummy_factors = FALSE)
   }
 
   # Partial
-  if (partial) {
-    data[[x]] <- effectsize::adjust(data[names(data) != y], multilevel = multilevel, bayesian = partial_bayesian)[[x]]
-    data[[y]] <- effectsize::adjust(data[names(data) != x], multilevel = multilevel, bayesian = partial_bayesian)[[y]]
+  if (!isFALSE(partial)) {
+    # partial
+    if (isTRUE(partial)) {
+      data[[x]] <- datawizard::adjust(data[names(data) != y], multilevel = multilevel, bayesian = partial_bayesian)[[x]]
+      data[[y]] <- datawizard::adjust(data[names(data) != x], multilevel = multilevel, bayesian = partial_bayesian)[[y]]
+    }
+
+    # semi-partial
+    if (partial == "semi") {
+      stop("Semi-partial correlations are not supported yet. Get in touch if you want to contribute.")
+    }
   }
 
   # Winsorize
@@ -170,20 +175,25 @@ cor_test <- function(data,
     }
 
     # winsorization would otherwise fail in case of NAs present
-    # data <- stats::na.omit(data)
-
-    data <- as.data.frame(winsorize(stats::na.omit(data[c(x, y)]), threshold = winsorize, verbose = verbose))
+    data <- as.data.frame(
+      datawizard::winsorize(stats::na.omit(data[c(x, y)]),
+        threshold = winsorize,
+        verbose = verbose
+      )
+    )
   }
 
   # Rank transform (i.e., "robust")
   if (ranktransform) {
-    data[c(x, y)] <- effectsize::ranktransform(data[c(x, y)], sign = FALSE, method = "average")
+    data[c(x, y)] <- datawizard::ranktransform(data[c(x, y)], sign = FALSE, method = "average")
   }
 
+  # check if enough no. of obs ------------------------------
+
+  # this is a trick in case the number of valid observations is lower than 3
   n_obs <- length(.complete_variable_x(data, x, y))
-  # This is a trick in case the number of valid observations is lower than 3
   invalid <- FALSE
-  if (n_obs < 3) {
+  if (n_obs < 3L) {
     if (isTRUE(verbose)) {
       warning(paste(x, "and", y, "have less than 3 complete observations. Returning NA."), call. = FALSE)
     }
@@ -288,7 +298,8 @@ cor_test <- function(data,
   # Output
   attr(out, "coefficient_name") <- c("rho", "r", "tau", "Dxy")[c("rho", "r", "tau", "Dxy") %in% names(out)][1]
   attr(out, "ci") <- ci
-  class(out) <- unique(c("easycorrelation", "parameters_model", class(out)))
+  attr(out, "data") <- data
+  class(out) <- unique(c("easycor_test", "easycorrelation", "parameters_model", class(out)))
   out
 }